refactor: board 도메인 패키지 구조
