+++ src/main/java/com/hanghae/prevstudy/domain/board/dto/BoardUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.hanghae.prevstudy.domain.board;
+package com.hanghae.prevstudy.domain.board.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -6,13 +6,9 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class BoardAddRequest {
-
+public class BoardUpdateRequest {
     @NotBlank(message = "제목을 입력해 주세요.")
     private final String title;
-
-    @NotBlank(message = "작성자를 입력해 주세요.")
-    private final String writer;
 
     @NotBlank(message = "내용을 입력해 주세요.")
     private final String content;
