+++ src/main/java/com/hanghae/prevstudy/domain/board/entity/Board.java
@@ -1,4 +1,4 @@
-package com.hanghae.prevstudy.domain.board;
+package com.hanghae.prevstudy.domain.board.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
