+++ src/main/java/com/hanghae/prevstudy/domain/board/repository/BoardRepository.java
@@ -1,5 +1,6 @@
-package com.hanghae.prevstudy.domain.board;
+package com.hanghae.prevstudy.domain.board.repository;
 
+import com.hanghae.prevstudy.domain.board.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
