+++ src/test/java/com/hanghae/prevstudy/domain/board/BoardServiceTest.java
@@ -1,6 +1,12 @@
 package com.hanghae.prevstudy.domain.board;
 
-import com.hanghae.prevstudy.global.exception.BoardErrorCode;
+import com.hanghae.prevstudy.domain.board.dto.BoardAddRequest;
+import com.hanghae.prevstudy.domain.board.dto.BoardResponse;
+import com.hanghae.prevstudy.domain.board.dto.BoardUpdateRequest;
+import com.hanghae.prevstudy.domain.board.entity.Board;
+import com.hanghae.prevstudy.domain.board.repository.BoardRepository;
+import com.hanghae.prevstudy.domain.board.service.BoardServiceImpl;
+import com.hanghae.prevstudy.domain.board.exception.BoardErrorCode;
 import com.hanghae.prevstudy.global.exception.PrevStudyException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
